Fixed loading state
