--- products/logs/frontend/sparkline-loading.scss
@@ -1,41 +1,0 @@
-.sparkline-loading {
-    position: relative;
-    opacity: 0.25;
-}
-
-.sparkline-loading-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-}
-
-.sparkline-loading-overlay::after {
-    position: absolute;
-    top: -50%; /* Adjust if shine is cut off due to skew */
-    left: -75%; /* Start position for animation, adjusted for skew */
-    width: 50%; /* Make the shine itself narrower */
-    height: 200%; /* Make it taller to cover diagonally */
-    content: '';
-    background: linear-gradient(
-        to right,
-        /* For skewed, a simple L-R gradient often works best before skewing */ transparent 0%,
-        rgb(255 255 255 / 60%) 50%,
-        transparent 100%
-    );
-    opacity: 0.8; /* Slightly more opaque shine */
-    transform: skewX(-25deg); /* Apply the skew */
-    animation: Sparkline__loading-shine-skewed 1.5s linear infinite; /* Faster, linear for constant speed */
-}
-
-@keyframes Sparkline__loading-shine-skewed {
-    0% {
-        left: -75%; /* Start far left */
-    }
-
-    100% {
-        left: 125%; /* End far right, ensures full sweep with skew */
-    }
-}
